delete old component
