--- vuu-ui/packages/vuu-filters/src/column-filter/index.ts
@@ -1,1 +1,0 @@
-export * from "./ColumnFilter";
