+++ vuu-ui/packages/vuu-filters/src/filter-utils.ts
@@ -1,6 +1,5 @@
 import {
   ColumnDescriptor,
-  KeyedColumnDescriptor,
 } from "@finos/vuu-datagrid-types";
 import {
   AndFilter,
@@ -10,7 +9,6 @@ import {
 } from "@finos/vuu-filter-types";
 import {
   extractFilterForColumn,
-  filterAsQuery,
   isAndFilter,
   isInFilter,
   isMultiClauseFilter,
@@ -19,6 +17,7 @@ import {
   isSingleValueFilter,
   partition,
 } from "@finos/vuu-utils";
+import { NumericOperator } from "./operator-utils";
 
 export const AND = "and";
 export const EQUALS = "=";
@@ -372,4 +371,40 @@ export const updateFilter = (
     return newFilter;
   }
   return filter;
+};
+
+export const getTypeaheadFilter = (
+  column: string,
+  filterValues: string[],
+  isStartsWithFilter?: boolean
+): Filter | undefined => {
+  if (filterValues.length === 0) {
+    return undefined;
+  }
+
+  if (isStartsWithFilter) {
+    // multiple starts with filters not currently supported
+    const startsWith = filterValues[0].substring(0, filterValues[0].length - 3);
+    return {
+      column,
+      op: "starts",
+      value: `"${startsWith}"`,
+    };
+  }
+
+  return {
+    column,
+    op: "in",
+    values: filterValues.map((value) => `"${value}"`),
+  };
+};
+
+export const getNumericFilter = (
+  column: string,
+  operator?: NumericOperator,
+  value?: number
+): Filter | undefined => {
+  if (operator === undefined) return undefined;
+  if (value === undefined || isNaN(value)) return undefined;
+  return { column, op: operator, value };
 };
